pyproject.toml
[tool.poetry]
name = "fastapiproject"
version = "0.1.0"
description = ""
authors = ["a-yermakova <ertya0909@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["standard"], version = "^0.115.3"}
uvicorn = "^0.32.0"
sqlalchemy = "^2.0.36"
asyncpg = "^0.30.0"
pydantic = "^2.9.2"
alembic = "^1.13.3"
passlib = "^1.7.4"
python-jose = "^3.3.0"
psycopg2 = "^2.9.10"
bcrypt = "^4.2.0"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.2"}
aioredis = "^2.0.1"
jinja2 = "^3.1.4"
aiogram = "^3.13.1"
celery = "^5.4.0"
gunicorn = "^23.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
